файлы js изменены на ts. Иправлены все ошибки
