+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+      "strict": true, // Включает все жесткие проверки типов
+      "outDir": "./dist/",
+      "noImplicitAny": false,
+      "module": "es6",
+      "target": "es5",
+      "allowJs": true,
+      "moduleResolution": "node"
+    }
+  }
